requirements.txt
Django==3.2.3
psycopg2-binary==2.8.6
django-statici18n==2.0.1
djangorestframework==3.12.4
python-dotenv==0.18.0
requests==2.25.1
celery==5.1.2
django-celery-beat==2.2.1
django-celery-results==2.2.0
factory-boy==3.2.0